requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
openai==0.28.0
PyPDF2==3.0.1
python-docx==1.0.1
openpyxl==3.1.2
pandas==2.1.1
numpy==1.26.0
requests==2.31.0
psycopg2-binary==2.9.9
Werkzeug==3.0.1
SQLAlchemy==2.0.23
gunicorn==21.2.0
cryptography==41.0.5
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
blinker==1.6.2
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
click==8.1.7
opencv-python==4.10.0.24
